feat: support scoped npm packages

+++ projects/npm-tools/packages/npm-scripts/src/utils/createTempFile.js
@@ -16,12 +16,16 @@ const BUILD_CONFIG = getMergedConfig('npmscripts', 'build');
  *
  * @param {string} filename the name of the file
  * @param {string|Buffer} content content of the file
+ * @param {object} options
+ * Pass {autoDelete:false} to prevent the file from being removed at exit.
+ *
+ * The default is {autoDelete:true}.
  *
  * @return {object}
  * An object with the `dispose` function (in case the caller wants to invoke it
  * before the end of the process) and the `filePath`.
  */
-function createTempFile(filename, content) {
+function createTempFile(filename, content, {autoDelete = true} = {}) {
 	const tempDirPath = path.join(BUILD_CONFIG.temp, 'tmp');
 
 	fs.ensureDirSync(tempDirPath);
@@ -31,7 +35,9 @@ function createTempFile(filename, content) {
 	fs.writeFileSync(tempFilePath, content);
 
 	const {dispose} = SignalHandler.onExit(() => {
-		fs.unlinkSync(tempFilePath);
+		if (autoDelete) {
+			fs.unlinkSync(tempFilePath);
+		}
 	});
 
 	return {
